Logit Normal is now the default timestep sampling technique for Flow Matching, following SOTA. It biases the samples to come from intermediate regions of the [0,1] set. Added denoising package to layers with all related diffusion/FM functionalities that don't depend on policy logic

+++ src/versatil/models/decoding/algorithm/flow_matching.py
@@ -116,7 +116,7 @@ class FlowMatching(DecodingAlgorithm):
                     logit_mean=self.logit_mean,
                     logit_std=self.logit_std,
                 )
-            epsilon = torch.randn_like(action)
+            epsilon = torch.randn_like(action.float())
             x_t = self.flow_matcher.sample_xt(x0=noise, x1=action, t=times, epsilon=epsilon)
             u_t = self.flow_matcher.compute_conditional_flow(x0=noise, x1=action, t=times, xt=x_t)
             interpolated_actions[key] = x_t
